move minigameStart from Main to Minigame class
